Extract filter code to separate module
